\name{BwQuant-package}
\alias{BwQuant-package}
\alias{BwQuant}
\docType{package}
\title{Bandwidth selectors for local linear quantile regression}
\description{The R package \code{BwQuant} implements different bandwidth selectors for local linear quantile regression, including selectors based on rule of thumb, plug-in and cross-validation tecniques.}
\author{Mercedes Conde-Amboage <mercedes.amboage@usc.es> and Cesar Sanchez-Sellero <cesar.sanchez@usc.es>

Maintainer: Mercedes Conde-Amboage <mercedes.amboage@usc.es>
}
\keyword{ package }